Update some enums
